requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
urllib3==1.26.6
Werkzeug==2.0.1